Added an undo button (right click)
Right-clicking anywhere will undo the last move made. It's a little buggy but works for the most part. Occasionally it will remove something totally different or crash, and I'm not sure why it does that but it definitely has something to do with not popping the move-history stack properly somewhere.